More signed/unsigned work
git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@34207 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ link-grammar/pp_lexer.c
@@ -1220,7 +1220,7 @@ int pp_lexer_count_commas_of_label(PPLexTable *lt)
   return n;
 }
 
-const char **pp_lexer_get_next_group_of_tokens_of_label(PPLexTable *lt, int *n_tokens)
+const char **pp_lexer_get_next_group_of_tokens_of_label(PPLexTable *lt, size_t *n_tokens)
 { 
   /* all tokens until next comma, null-terminated */
   int n;
